Make library usable on Pythons without GUI libraries
Some Python distributions (notably, the one used on Heroku) are not
built to include the GUI libraries (`tkinter`). We'd like to use
Turtlethread in a web server (https://hedycode.com) but are running
into issues because of this.

Even though `turtlethread` doesn't really need the GUI libraries to do
its main job, just the presence of the `visualize.py` library and it
containing a `from turtle import ...` statement, and the built-in
`turtle` library depending on `tkinter`, makes importing `turtlethread`
fail:

```
>>> import turtlethread
...
ModuleNotFoundError: No module named '_tkinter'
```

This PR moves the import of the built-in turtle module *into* the
`visualize_pattern()` function, so that we'll only fail if the
function is actually called; not when the module is imported.